backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pyjwt==2.9.0
pycryptodome==3.20.0
cryptography==43.0.1
bcrypt==4.2.0
uuid==1.30
gunicorn==23.0.0